+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # CHANGELOG
 
+## 12.1.33
+- When dropping a Specialism skill on to the character sheet where the specialism hasn't been chosen you will get asked for the specialism name
+- When the character sheet is unlocked the Skills Tab context menu now includes a "Recalculate skill base scores" which does exactly that.
+- When dropping Personality on to a character - the skills included in the personality are added to the character - with relevant choices presented to the player and base scores calculated. 
+- The same applies for a Profession
+- Skills added via the Personality gain the +20 bonus and are highlighted in the Skill Dev tab (Skill tab when character sheet unlocked), though you can edit the Personality Points on any skill
+- Skills added via the Professoin are highlighted in the Skill Dev tab, though the system doesn't restrict you to spending Professional Skill Points on just these skill
+- NPCs now have a Power Modifier option on the Base Stats tab
+- NPCs now have a 'Move Notes" entry on the base stats tab for a longer description of their movement.  This appears as a tooltip on the main NPC sheet
+- GMs should now be able to drag all items from the Characters and NPCs to the items menu or compendia
+- The skill category on the Skill Item Sheet should now display properly for players (it was fine for GMs)
+
 ## 12.1.32
 - Fixed issue with the Sorcery Spell sheet nullifying levels when opened by a player
 - Fixed issue with Psychic abilities not being able to access Description/GM Notes